Move test to deprecated_test and improve message
